fixed the :nil bug of tweet_helper)
Fixed the bug in the tweets helper where it would throw a :nil error as the tweet id not found. The error was because of the memoization that I was doing on the current_user_liked_this_tweet function. It was not running the query again to check if the user has liked this tweet and because of that it was calling the delete tweet path, which caused it to fail